ai_modules/faiss_search-main/requirements.txt
fastapi
uvicorn
requests
faiss-cpu
pymysql
numpy
pydantic
dotenv
openai
cryptography